FRONTEND: Modularización del Sistema Selnic, antes llamado DMS, terminando estructura modular por carpeta, proyecto ejecutandose
