+++ frontend/src/App.jsx
@@ -8,26 +8,27 @@ import AuthLayout from "./layout/Auth.layout";
 import RutaProtegida from "./layout/RutaProtegida.layout";
 import RutaProtegidaCoodinador from "./layout/RutaProtegidaCoordinador.layout";
 
-import Login from "./pages/auth/Login.pages";
-import ConfirmarCuenta from "./pages/auth/ConfirmarCuenta.pages";
-import Registrar from "./pages/auth/Registrar.pages";
-import OlvidePassword from "./pages/auth/OlvidePassword.pages";
-import NuevoPassword from "./pages/auth/NuevoPassword.pages";
-import AdministrarTramites from "./pages/admin/AdministrarTramites.pages";
-import TramitesAsignados from "./pages/admin/TramitesAsignados.pages";
-import TramitesAsignarReasignar from "./pages/admin/TramitesAsignarReasignar.pages";
-import TramitesCompletados from "./pages/admin/TramitesCompletados.pages";
-import ConsultarTramites from "./pages/admin/ConsultarTramites.pages";
-import EditarPerfil from "./pages/admin/EditarPerfil.pages";
-import CambiarPassword from "./pages/admin/CambiarPassword.pages";
-import TramitesPorDespachar from "./pages/admin/TramitesPorDespachar.pages";
+import Login from "./modules/administration/pages/auth/Login.pages";
+import ConfirmarCuenta from "./modules/administration/pages/auth/ConfirmarCuenta.pages";
+import Registrar from "./modules/administration/pages/auth/Registrar.pages";
+import OlvidePassword from "./modules/administration/pages/auth/OlvidePassword.pages";
+import NuevoPassword from "./modules/administration/pages/auth/NuevoPassword.pages";
+import CambiarPassword from "./modules/administration/pages/auth/CambiarPassword.pages";
+import EditarPerfil from "./modules/administration/pages/auth/EditarPerfil.pages";
+import AdminDMS from "./modules/administration/pages/admin-panel/AdminDMS.pages";
+import Empleados from "./modules/administration/pages/admin-panel/Empleados.pages";
+import Departamentos from "./modules/administration/pages/admin-panel/Departamentos.pages";
 
-import AdminDMS from "./pages/admin/AdminDMS.pages";
-import Empleados from "./pages/admin/pagesConfiguracion/Empleados.pages";
-import Departamentos from "./pages/admin/pagesConfiguracion/Departamentos.pages";
-import Tramites from "./pages/admin/pagesConfiguracion/Tramites.pages";
-import InicioDMS from "./pages/admin/pagesConfiguracion/InicioDMS.pages";
-import InicioDMSUsuario from "./pages/admin/InicioDMSUsuario.pages";
+import AdministrarTramites from "./modules/document-management/pages/AdministrarTramites.pages";
+import TramitesAsignados from "./modules/document-management/pages/TramitesAsignados.pages";
+import TramitesAsignarReasignar from "./modules/document-management/pages/TramitesAsignarReasignar.pages";
+import TramitesCompletados from "./modules/document-management/pages/TramitesCompletados.pages";
+import ConsultarTramites from "./modules/document-management/pages/ConsultarTramites.pages";
+import TramitesPorDespachar from "./modules/document-management/pages/TramitesPorDespachar.pages";
+
+import Tramites from "./modules/administration/pages/admin-panel/Tramites.pages";
+import InicioDMS from "./modules/administration/pages/admin-panel/InicioDMS.pages";
+import InicioDMSUsuario from "./modules/administration/pages/admin-panel/InicioDMSUsuario.pages";
 
 function App() {
   return (
