+++ frontend/src/components/Header.components.jsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import useAuth from "../../../hooks/useAuth.hook";
+import useAuth from "../hooks/useAuth.hook";
 
 const Header = () => {
   const { auth, cerrarSesion } = useAuth();
