+++ frontend/src/modules/document-management/pages/TramitesPorDespachar.pages.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import TablaTramitesBusqueda from "../../../components/TablaTramitesBusqueda.components";
+import TablaTramitesBusqueda from "../components/TablaTramitesBusqueda.components";
 import useTramites from "../../../hooks/useTramites.hook";
 
 const TramitesPorDespachar = () => {
